add custom layer for topk
